fp32 fp16-GEMM path: run GELU directly on fp16, no fp32 round-trip
ATen's GELU kernel already accumulates internally in fp32 regardless of
input dtype (same pattern already found for softmax and LayerNorm), so
F.gelu(fp16_tensor) is bit-identical (torch.equal, not just close) to
upcasting to fp32, computing gelu, and casting back to fp16 -- verified
directly and via the full 6-layer model at 5 configs including
input-scale=0.5. The previous code paid for that round-trip anyway (2
elementwise kernel launches, full fp32 memory traffic for a tensor that's
immediately cast back down).

default_fp32: 1.586x -> 1.786x vs the compiled bar (1.528ms -> 1.358ms).
Geomean 1.060x -> 1.072x. 5/5 accuracy PASS, max_abs unchanged at every
tested shape (bit-exact continuation, not a new approximation). fp16/bf16
paths untouched (they don't use this GEMM path).

+++ torch_transformer_benchmark.py
@@ -1129,9 +1129,11 @@ class UserOptimizedTransformer(BaselineTransformer):
         pieces = []
         for start in range(0, flat.shape[0], chunk_size):
             piece = flat[start : start + chunk_size]
-            hidden = F.linear(piece.to(torch.float16), ffn_in_weight16, ffn_in_bias16).to(torch.float32)
+            # No fp32 round-trip around GELU -- see the non-chunked branch's
+            # comment; bit-identical, verified torch.equal.
+            hidden = F.linear(piece.to(torch.float16), ffn_in_weight16, ffn_in_bias16)
             hidden = F.gelu(hidden, approximate="none")
-            pieces.append(F.linear(hidden.to(torch.float16), ffn_out_weight16, ffn_out_bias16).to(torch.float32))
+            pieces.append(F.linear(hidden, ffn_out_weight16, ffn_out_bias16).to(torch.float32))
         return torch.cat(pieces, dim=0).reshape(batch, seq_len, d_model)
 
     def _forward_core(
@@ -1223,9 +1225,16 @@ class UserOptimizedTransformer(BaselineTransformer):
                 ffn_out_weight16, ffn_out_bias16 = _get_linear_fp16_weights(layer.ffn_out)
                 chunk_size = self._ffn_chunk_size(batch, seq_len, self.config.ffn_dim)
                 if chunk_size is None:
-                    hidden = F.linear(normed2.to(torch.float16), ffn_in_weight16, ffn_in_bias16).to(torch.float32)
+                    # GELU runs directly on the fp16 GEMM output, no upcast
+                    # round-trip: F.gelu on a half tensor is bit-identical to
+                    # upcast->gelu->downcast (ATen's GELU kernel already
+                    # accumulates in fp32 internally regardless of input
+                    # dtype, same as softmax/LayerNorm) -- verified
+                    # torch.equal, not just close. Removes 2 elementwise
+                    # kernel launches and halves this op's memory traffic.
+                    hidden = F.linear(normed2.to(torch.float16), ffn_in_weight16, ffn_in_bias16)
                     hidden = F.gelu(hidden, approximate="none")
-                    ffn_out = F.linear(hidden.to(torch.float16), ffn_out_weight16, ffn_out_bias16).to(torch.float32)
+                    ffn_out = F.linear(hidden, ffn_out_weight16, ffn_out_bias16).to(torch.float32)
                 else:
                     ffn_out = self._chunked_ffn_fp16gemm(
                         normed2, ffn_in_weight16, ffn_in_bias16, ffn_out_weight16, ffn_out_bias16, chunk_size
